DAOS-14010 object: revert size query patch to retry
revert 13571 to see if ec test pass.

Test-tag: ec
Required-githooks: true
