+++ src/object/obj_ec.h
@@ -1,5 +1,5 @@
 /**
- * (C) Copyright 2019-2024 Intel Corporation.
+ * (C) Copyright 2019-2023 Intel Corporation.
  *
  * SPDX-License-Identifier: BSD-2-Clause-Patent
  */
@@ -317,11 +317,6 @@ struct obj_reasb_req;
 #define obj_ec_shard_off(obj, dkey_hash, shard)				\
 	((shard % obj_ec_tgt_nr(&obj->cob_oca) + obj_ec_tgt_nr(&obj->cob_oca) -		\
 	 obj_ec_shard_idx(obj, dkey_hash, 0)) % obj_ec_tgt_nr(&obj->cob_oca))
-
-/* Get the logical offset of shard within one group by oca, physical idx -> logical idx */
-#define obj_ec_shard_off_by_oca(layout_ver, dkey_hash, oca, shard)				\
-	((shard % obj_ec_tgt_nr(oca) + obj_ec_tgt_nr(oca) -					\
-	 obj_ec_shard_idx_by_layout_ver(layout_ver, dkey_hash, oca, 0)) % obj_ec_tgt_nr(oca))
 
 /* Get the logical offset of the tgt_idx by start target of EC, physical idx -> logical idx */
 #define obj_ec_shard_off_by_start(tgt_idx, oca, start_tgt)		\
