+++ src/tests/suite/daos_obj.c
@@ -1,5 +1,5 @@
 /**
- * (C) Copyright 2016-2024 Intel Corporation.
+ * (C) Copyright 2016-2023 Intel Corporation.
  *
  * SPDX-License-Identifier: BSD-2-Clause-Patent
  */
@@ -5191,124 +5191,6 @@ io_52(void **state)
 	obj_coll_punch(arg, OC_EC_4P1GX);
 }
 
-static void
-obj_coll_query(test_arg_t *arg, daos_oclass_id_t oclass)
-{
-	daos_obj_id_t	oid;
-	daos_handle_t	oh;
-	daos_iod_t	iod = { 0 };
-	d_sg_list_t	sgl = { 0 };
-	daos_recx_t	recx = { 0 };
-	d_iov_t		val_iov;
-	d_iov_t		dkey;
-	d_iov_t		akey;
-	uint64_t	dkey_val;
-	uint64_t	akey_val;
-	uint32_t	update_var = 0xdeadbeef;
-	uint32_t	flags;
-	int		rc;
-
-	/** init dkey, akey */
-	dkey_val = akey_val = 0;
-	d_iov_set(&dkey, &dkey_val, sizeof(uint64_t));
-	d_iov_set(&akey, &akey_val, sizeof(uint64_t));
-
-	oid = daos_test_oid_gen(arg->coh, oclass, DAOS_OT_MULTI_UINT64, 0, arg->myrank);
-	rc = daos_obj_open(arg->coh, oid, DAOS_OO_RW, &oh, NULL);
-	assert_rc_equal(rc, 0);
-
-	dkey_val = 5;
-	akey_val = 10;
-	iod.iod_type = DAOS_IOD_ARRAY;
-	iod.iod_name = akey;
-	iod.iod_recxs = &recx;
-	iod.iod_nr = 1;
-	iod.iod_size = sizeof(update_var);
-
-	d_iov_set(&val_iov, &update_var, sizeof(update_var));
-	sgl.sg_iovs = &val_iov;
-	sgl.sg_nr = 1;
-
-	recx.rx_idx = 5;
-	recx.rx_nr = 1;
-
-	rc = daos_obj_update(oh, DAOS_TX_NONE, 0, &dkey, 1, &iod, &sgl, NULL);
-	assert_rc_equal(rc, 0);
-
-	dkey_val = 10;
-	val_iov.iov_buf_len += 1024;
-	rc = daos_obj_update(oh, DAOS_TX_NONE, 0, &dkey, 1, &iod, &sgl, NULL);
-	assert_rc_equal(rc, 0);
-	d_iov_set(&val_iov, &update_var, sizeof(update_var));
-
-	recx.rx_idx = 50;
-	rc = daos_obj_update(oh, DAOS_TX_NONE, 0, &dkey, 1, &iod, &sgl, NULL);
-	assert_rc_equal(rc, 0);
-
-	flags = DAOS_GET_DKEY | DAOS_GET_AKEY | DAOS_GET_RECX | DAOS_GET_MAX;
-	rc = daos_obj_query_key(oh, DAOS_TX_NONE, flags, &dkey, &akey, &recx, NULL);
-	assert_rc_equal(rc, 0);
-	assert_int_equal(*(uint64_t *)dkey.iov_buf, 10);
-	assert_int_equal(*(uint64_t *)akey.iov_buf, 10);
-	assert_int_equal(recx.rx_idx, 50);
-	assert_int_equal(recx.rx_nr, 1);
-
-	flags = DAOS_GET_AKEY | DAOS_GET_RECX | DAOS_GET_MAX;
-	rc = daos_obj_query_key(oh, DAOS_TX_NONE, flags, &dkey, &akey, &recx, NULL);
-	assert_rc_equal(rc, 0);
-	assert_int_equal(*(uint64_t *)akey.iov_buf, 10);
-	assert_int_equal(recx.rx_idx, 50);
-	assert_int_equal(recx.rx_nr, 1);
-
-	flags = DAOS_GET_RECX | DAOS_GET_MAX;
-	rc = daos_obj_query_key(oh, DAOS_TX_NONE, flags, &dkey, &akey, &recx, NULL);
-	assert_rc_equal(rc, 0);
-	assert_int_equal(recx.rx_idx, 50);
-	assert_int_equal(recx.rx_nr, 1);
-
-	rc = daos_obj_close(oh, NULL);
-	assert_rc_equal(rc, 0);
-}
-
-static void
-io_53(void **state)
-{
-	test_arg_t	*arg = *state;
-
-	print_message("Collective object query - OC_SX\n");
-
-	if (!test_runable(arg, 2))
-		return;
-
-	obj_coll_query(arg, OC_SX);
-}
-
-static void
-io_54(void **state)
-{
-	test_arg_t	*arg = *state;
-
-	print_message("Collective object query - OC_EC_2P1G2\n");
-
-	if (!test_runable(arg, 3))
-		return;
-
-	obj_coll_query(arg, OC_EC_2P1G2);
-}
-
-static void
-io_55(void **state)
-{
-	test_arg_t	*arg = *state;
-
-	print_message("Collective object query - OC_EC_4P1GX\n");
-
-	if (!test_runable(arg, 5))
-		return;
-
-	obj_coll_query(arg, OC_EC_4P1GX);
-}
-
 static const struct CMUnitTest io_tests[] = {
 	{ "IO1: simple update/fetch/verify",
 	  io_simple, async_disable, test_case_teardown},
@@ -5413,12 +5295,6 @@ static const struct CMUnitTest io_tests[] = {
 	  io_51, NULL, test_case_teardown},
 	{ "IO52: collective punch object - OC_EC_4P1GX",
 	  io_52, NULL, test_case_teardown},
-	{ "IO53: collective object query - OC_SX",
-	  io_53, async_disable, test_case_teardown},
-	{ "IO54: collective object query - OC_EC_2P1G2",
-	  io_54, async_disable, test_case_teardown},
-	{ "IO55: collective object query - OC_EC_4P1GX",
-	  io_55, async_disable, test_case_teardown},
 };
 
 int
